Move insetting code to extension
